Split some utilities out of config
To be used in other config instances
