+++ src/utils.h
@@ -11,6 +11,8 @@ struct Utils {
   static std::string fromBase64(std::string);
   static std::string toBase64(const std::string &);
   static void hex2bin(const std::string hexstring, unsigned char *binout);
+  static std::string stripQuotes(const std::string &value);
+  static std::string addQuotes(const std::string &value);
   static Json::Value parseJSON(const std::string &json_str);
   static Json::Value parseJSONFile(const boost::filesystem::path &filename);
   static std::string genPrettyName();
